port 68510-68511,68517 from trunk
git-svn-id: https://svn.r-project.org/R/branches/R-3-2-branch@68534 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/envir.c
@@ -2950,7 +2950,7 @@ int envlength(SEXP rho)
     if(IS_USER_DATABASE(rho)) {
         R_ObjectTable *tb = (R_ObjectTable*)
 	    R_ExternalPtrAddr(HASHTAB(rho));
-        return(xlength(tb->objects(tb)));
+        return (int) xlength(tb->objects(tb));
     } else if( HASHTAB(rho) != R_NilValue)
         return HashTableSize(HASHTAB(rho), 1);       
     else
